Support light source instances
Also nested instanced. It should works. But the algorithm is too complicated. It requires to store three additional structures: 
1. From each top level instance to real objects
2. From each instance to all top level instance, where it is subinstance (actual nested instance)
3. From real object to all top level instance, contained this object

We create these structures at the scene creation process, and then use them when we should update transforms of an real object (use 3-rd structure), or update transfrom of an instance object (use 1-st and 2-nd structures).