Support instances from pointclouds
And also support primitive particle shapes: cube, sphere and so on. Override only color of the instance object. All other attributes are not synced. Because if master is real scene object, then it can used in non-particle context. One primitive shape can be used from different pointcloud, so, contains different attributes.

But may be primitive shapes should be exported per pointcloud. Then we can use pointcloud attributes for it.