Implement series rendering
This mode can be turned on on the config file. We does not use UI, because this mode very technical and should not used by any user. This mode allows to save intermediate render results. The main purpose of these images is to use it for training denoising neural networks.

The system save intermidiate results with any sample step. But if the rendering device is very powerfull, then it can jump over several steps duiring one render update cycle. It can be fixed by setting fixed number of used cores for the cpu.