Support denoising
There are some problems with denoising and light groups. In Blender all lightgroups saved without denoising (eve it is active). But in our cases, if we use denoising, then the render is crashed when it clear the session.

So, we check active denoising and switch off all lightgroup passes. From output and also from visual.