Disable reassign color when update geometry
Fix #1. Now color, name and asset name of the object assigned only when the scene is created at the first time. These values does not changed during updates. So, if we change object color, then we should force to recreate the scene, because render will use the old value.

But object colors are not used in most cases. Only poincloud instances use it. There colors updated properly, because when we change poincloud atribute it recreate the scene from scratch. It's hard to find all instences, generated by poincloud.