Implement force to create session from scratch
Some enhancement for #4.
Introduce render parameter "Update Mode". This is an enum with two values (in this implementation): always abort, update and abort. If the first item is selected, then the session creates from scratch every time. In this mode no update internal data is used. If the second item is selected, then it tries to update all scene items, and if something wron,g, then abort and start clear session.