Once more improvement ofa  strand size rendering
Remove tip_prop property for pointcloud. By default each strand is rendered with the size of initial point. If we would like to decresase the size during the strand, then StrandSize attribute should be used. It's enough to create 2-valued array for each strand, and then these two values will be define the profile of the strand. 

Additional compound Set Strand Profile is added. It allows to contol the strand profile by using f-curve.