Implement #7, supports gpu rendering for shaderballs
Add parameter use_gpu to the config.ini file. By defaul it set to 0. It means that cpu device should be used for rendering shaderball previes. If the parameter is set to 1, then it tries to use gpu. But for now osl rendering does not support by gpu. So, is osl is activated (used_osl = 1), then it disables the gpu and alwasy use cpu.