Implement #28
Allows to load velocity grids with custom names for rendering vdbs.

There is a bug with vdb rendering. When we change any parameter of the VDBPrimitive, then the volume is skiped from the rendereing. Even when we reload all grids. It looks like something with caching in the Cycles side. After full rebuild of the scene, all renders correctly.