Initial implementation of curve and surface rendering
By default curves and surfaces are not rendered. To render curve it's required to add the property Cycles Curve. For surfaces - CyclesSurface. The main parameter in both properties - the number of samples. Each curve converted to the strand (hair promitive in Cycles). Each surface renders as square patch. The umber of u- and v-samples define the polygon density.

In Softimage curves are not rendered, so, it's impossible to assgn the material to the curve. To render it, the property CyclesCurve ontains parameter which allows to select material from the scene.

Each curve support following attributes:
* Intercept
* Random
* Length

Each surface support attributes:
* Random per island
* UV, which are procedural and fill the square [0; 1]x[0; 1]. Each subsurface maps to this square

Surface trims are not supported for now.